Added $action parameter to process_actions()
Added a parameter to the process_actions() function so that the functionality can be used outside of the UI interface. For example, it can be called from the save_post action to delete transients with expirations (to clear cached queries, for instance).

Note: I'm don't generally do pull requests, so I hope I'm doing this right, heh.